+++ math_utils/lin_sin.c
@@ -7,8 +7,8 @@
 #define LOW_POINT -(1<<(SF_ATAN_OUT-1))  // common points of approximation are pi/2 and -pi/2, scaled
 #define HIGH_POINT (1<<(SF_ATAN_OUT-1))
 
-int lin_sin(int theta) {
-    int result;    
+int16_t lin_sin(int16_t theta) {
+    int16_t result;    
     
     if (theta < LOW_POINT) {
         result = -SLOPE_SINCOS * theta - Y_INT;
